requirements.txt
telebot==0.0.5
pandas==1.5.3
matplotlib==3.7.2
Pillow==9.5.0
torch==1.13.1
xgboost==1.6.2  
lightgbm==3.3.2
joblib==1.1.0